Renamespace source compilation events
