+++ src/Event/SourceCompilationStartedEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Event\SourceCompilation;
+namespace App\Event;
 
 class SourceCompilationStartedEvent extends AbstractSourceEvent
 {
