+++ src/Event/SourceEventInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Event\SourceCompilation;
+namespace App\Event;
 
 use Psr\EventDispatcher\StoppableEventInterface;
