+++ src/Services/TestFactory.php
@@ -6,7 +6,7 @@ namespace App\Services;
 
 use App\Entity\Test;
 use App\Entity\TestConfiguration;
-use App\Event\SourceCompilation\SourceCompilationPassedEvent;
+use App\Event\SourceCompilationPassedEvent;
 use App\Repository\TestRepository;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use webignition\BasilCompilerModels\TestManifest;
