+++ src/Services/WorkerEventFactory/EventHandler/AbstractEventHandler.php
@@ -14,10 +14,10 @@ use App\Event\JobCompletedEvent;
 use App\Event\JobFailedEvent;
 use App\Event\JobReadyEvent;
 use App\Event\JobTimeoutEvent;
-use App\Event\SourceCompilation\SourceCompilationFailedEvent as CompilationFailedEvent;
-use App\Event\SourceCompilation\SourceCompilationPassedEvent as CompilationPassedEvent;
-use App\Event\SourceCompilation\SourceCompilationStartedEvent as CompilationStartedEvent;
-use App\Event\SourceCompilation\SourceEventInterface;
+use App\Event\SourceCompilationFailedEvent as CompilationFailedEvent;
+use App\Event\SourceCompilationPassedEvent as CompilationPassedEvent;
+use App\Event\SourceCompilationStartedEvent as CompilationStartedEvent;
+use App\Event\SourceEventInterface;
 use App\Event\StepEventInterface;
 use App\Event\StepFailedEvent;
 use App\Event\StepPassedEvent;
